requirements.txt
Django==1.3.1
django-grappelli==2.3.5
Fabric==1.3.3
django-compressor==1.1.2
pyscss==1.1.3
mimeparse>=0.1.3
django-tastypie==0.9.11
